+++ src/Layers/xrRender/r__sector.cpp
@@ -47,7 +47,7 @@ void CPortal::OnRender	()
 		RCache.set_xform_world(Fidentity);
 		// draw solid
 		RCache.set_Shader	(dxRenderDeviceRender::Instance().m_SelectionShader);
-		RCache.dbg_Draw		(D3DPT_TRIANGLEFAN,&*V.begin(),V.size()-2);
+		RCache.dbg_Draw		(D3DPT_TRIANGLEFAN,&*V.begin(), (int)V.size()-2);
 
 		// draw wire
 		if (bDebug){
@@ -56,7 +56,7 @@ void CPortal::OnRender	()
 			Device.SetNearer(TRUE);
 		}
 		RCache.set_Shader	(dxRenderDeviceRender::Instance().m_WireShader);
-		RCache.dbg_Draw		(D3DPT_LINESTRIP,&*(V.begin()+1),V.size()-2);
+		RCache.dbg_Draw		(D3DPT_LINESTRIP,&*(V.begin()+1), (int)V.size()-2);
 		if (bDebug){
 			RImplementation.rmNormal();
 		}else{
