+++ src/Layers/xrRenderDX10/dx10ConstantBuffer.cpp
@@ -70,7 +70,7 @@ bool dx10ConstantBuffer::Similar(dx10ConstantBuffer &_in)
 
 	VERIFY(m_MembersNames.size() == _in.m_MembersNames.size());
 
-	int iMemberNum = m_MembersNames.size();
+	int iMemberNum = (int)m_MembersNames.size();
 	for ( int i=0; i<iMemberNum; ++i)
 	{
 		if (m_MembersNames[i].c_str()!=_in.m_MembersNames[i].c_str())
