+++ src/Layers/xrRenderDX10/dx11HW.cpp
@@ -655,7 +655,7 @@ void fill_vid_mode_list(CHW* _hw)
 		_tmp.back()					= xr_strdup(str);
 	}
 	
-	u32 _cnt						= _tmp.size()+1;
+	u32 _cnt						= (u32)_tmp.size()+1;
 
 	vid_mode_token					= xr_alloc<xr_token>(_cnt);
