+++ src/Layers/xrRenderPC_R2/r2_loader.cpp
@@ -226,13 +226,13 @@ void CRender::LoadBuffers		(CStreamReader *base_fs,	BOOL _alternative)
 			fs->r				(dcl,buffer_size);
 			fs->advance			(-(int)buffer_size);
 
-			u32 dcl_len = GetDeclLength(dcl) + 1;
+			u32 dcl_len = u32(GetDeclLength(dcl) + 1);
 			_DC[i].resize		(dcl_len);
 			fs->r				(_DC[i].begin(),dcl_len*sizeof(D3DVERTEXELEMENT9));
 
 			// count, size
 			u32 vCount			= fs->r_u32	();
-			u32 vSize = ComputeVertexSize(dcl, 0);
+			u32 vSize = (u32)ComputeVertexSize(dcl, 0);
 #ifdef DEBUG
 			Msg("* [Loading VB] %d verts, %d Kb", vCount, (vCount * vSize) / 1024);
 #endif // DEBUG
